editors - more cleanup of large files
